py4seo/ht/dz12/p2/requirements.txt
Pillow==6.2.1
requests-html==0.10.0